[#5] shutdown 로직 함수로 분리
