+++ src/constants.rs
@@ -1,0 +1,2 @@
+pub const PROXY_HOST_HEADER: &str = "Proxy-Host";
+pub const UPGRADE_SOCKET_PATH: &str = "/tmp/affogato_upgrade.sock";
